requirements.txt
aiohttp>=3.8.0
asyncio>=3.4.3
beautifulsoup4>=4.11.0
requests>=2.28.0
colorama>=0.4.6
Jinja2>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
click>=8.1.0
lxml>=4.9.0
urllib3>=1.26.0
fake-useragent>=1.4.0
pyyaml>=6.0
chardet>=5.0.0
